pyproject.toml
[project]
name = "providerserver"
version = "0.1.0"
description = ""
authors = [
    {name = "avinash84319",email = "avihanvaisri@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "libvirt-python (>=11.0.0,<12.0.0)",
    "paramiko (>=3.5.1,<4.0.0)",
    "prometheus-client (>=0.21.1,<0.22.0)",
    "keyring (>=25.6.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
